Restore batched withdrawals support
